requirements.txt
backoff==1.11.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
Deprecated==1.2.13
googleapis-common-protos==1.56.2
grpcio==1.46.3
idna==3.3
importlib-metadata==4.11.4
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
opentelemetry-api==1.12.0rc1
opentelemetry-distro==0.31b0
opentelemetry-exporter-otlp==1.11.1
opentelemetry-exporter-otlp-proto-grpc==1.11.1
opentelemetry-exporter-otlp-proto-http==1.11.1
opentelemetry-instrumentation==0.31b0
opentelemetry-instrumentation-aws-lambda==0.31b0
opentelemetry-instrumentation-dbapi==0.31b0
opentelemetry-instrumentation-flask==0.31b0
opentelemetry-instrumentation-jinja2==0.31b0
opentelemetry-instrumentation-logging==0.31b0
opentelemetry-instrumentation-sqlite3==0.31b0
opentelemetry-instrumentation-urllib==0.31b0
opentelemetry-instrumentation-wsgi==0.31b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-proto==1.11.1
opentelemetry-sdk==1.12.0rc1
opentelemetry-semantic-conventions==0.31b0
opentelemetry-util-http==0.31b0
protobuf==3.20.1
requests==2.27.1
six==1.16.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
